back/functions/requirements.txt
firebase_functions~=0.1.0
openai==1.3.4
python-dotenv==1.0.0
./mystorytime
requests